remove null handling
we had shitload of code to handle NULL as istore value
while we where still not able to build consistent behavior
originally NULL was introduced to deal with missing keys in certain
operations
i.e.
SELECT '1=>1, 2=>2'::istore * '2=>2, 3=>3';
            ?column?
--------------------------------
 "1"=>NULL, "2"=>"4", "3"=>NULL

while
SELECT '1=>1, 2=>2'::istore + '2=>2, 3=>3';
           ?column?
------------------------------
 "1"=>"1", "2"=>"4", "3"=>"3"

while the latter behavior is quite useful for incremental operations
it's totally inconsistent with the former.

This patch keeps the inconsistent behavior but removes NULL value keys
from the output.

Which results in

SELECT '1=>1, 2=>2'::istore * '2=>2, 3=>3';
 ?column?
----------
 "2"=>"4"

and

SELECT '1=>1, 2=>2'::istore + '2=>2, 3=>3';
           ?column?
------------------------------
 "1"=>"1", "2"=>"4", "3"=>"3"

See spec diff for all the differences